+++ packages/btn-dropdown/src/BtnDropdown.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
+import { useTemplateRef } from 'vue';
+import { type ComponentExposed } from 'vue-component-type-helpers';
 import BtnDropdownSingle from './BtnDropdownSingle.vue';
 import BtnDropdownSplit from './BtnDropdownSplit.vue';
-import { BtnDropdownEvents, BtnDropdownProps } from './useDropdownHandler';
+import { BtnDropdownEvents, BtnDropdownProps, BtnDropdownSlotProps } from './useDropdownHandler';
 
 const props = withDefaults(defineProps<{
     split?: boolean
@@ -11,10 +13,26 @@ const props = withDefaults(defineProps<{
 });
 
 const emit = defineEmits<BtnDropdownEvents>();
+
+const el = useTemplateRef<ComponentExposed<typeof BtnDropdownSingle | typeof BtnDropdownSplit>>('el');
+
+defineExpose({
+    hide: () => el.value?.hide(),
+    show: () => el.value?.show(),
+    toggle: () => el.value?.toggle(),
+});
+
+defineSlots<{
+    button(props: BtnDropdownSlotProps): any;
+    toggle(props: BtnDropdownSlotProps): any;
+    split(props: BtnDropdownSlotProps): any;
+    default(): any;
+}>();
 </script>
 
 <template>
     <Component
+        ref="el"
         :is="!split ? BtnDropdownSingle : BtnDropdownSplit"
         v-bind="props"
         @click="(e: MouseEvent) => emit('click', e)"