+++ packages/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vue-interface/dropdown-menu
 
+## 3.0.7
+
+### Patch Changes
+
+- 7a408dc: Fixed typescript issues
+
 ## 3.0.6
 
 ### Patch Changes